blogapp/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.13
django-ckeditor==5.0.0
django-environ==0.11.2
django-js-asset==2.2.0
django-seo==0.3.5
exceptiongroup==1.2.2
frozenlist==1.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==0.28.0
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.0
urllib3==2.2.2
yarl==1.9.4
